Day 10 - Part 1
Beet my head against this longer than I would have liked, but only a few hours overall. I missunderstood and started trying to find every path to a 9 from a 0 and counting each of those paths, not just accessibility of the 9 from the 0, but the unique ways it could be accessed. The algorithm is still finding all the ways... I'm just set()ing them out at the end.