Day 11 - Part 2
Went down a very deep hole trying to get this one. Took a few days off for work/travel/vacation. read some stuff that lead to the idea that you have to cache it to even have a possiblity of finishing in a reasonable time.

So I learned about memoization and `functools.cache`!

Also included a version (prior to the memoization) where I tried to use the 'stackless' functionality in pypy, without the cache I don't have a good idea if it really helped anything or not cause it still never finished.